refactor component registry into multiple files
