+++ lightyear/src/shared/replication/receive.rs
@@ -6,7 +6,7 @@ use super::{EntityActionsMessage, EntityUpdatesMessage, SpawnAction};
 use crate::packet::message::MessageId;
 use crate::prelude::client::Confirmed;
 use crate::prelude::{ClientId, Tick};
-use crate::protocol::component::ComponentRegistry;
+use crate::protocol::component::registry::ComponentRegistry;
 use crate::serialize::reader::Reader;
 use crate::shared::events::connection::ConnectionEvents;
 use crate::shared::replication::authority::{AuthorityPeer, HasAuthority};
